+++ druid/const.lua
@@ -5,18 +5,18 @@
 
 local M = {}
 
-M.ACTION_TEXT = hash("text")
-M.ACTION_MARKED_TEXT = hash("marked_text")
+M.ACTION_TEXT = hash(sys.get_config("druid.input_text", "text"))
+M.ACTION_TOUCH = hash(sys.get_config("druid.input_touch", "touch"))
+M.ACTION_MARKED_TEXT = hash(sys.get_config("druid.input_marked_text", "marked_text"))
 
-M.ACTION_BACKSPACE = hash("key_backspace")
-M.ACTION_ENTER = hash("key_enter")
-M.ACTION_BACK = hash("key_back")
-M.ACTION_ESC = hash("key_esc")
+M.ACTION_ESC = hash(sys.get_config("druid.input_key_esc", "key_esc"))
+M.ACTION_BACK = hash(sys.get_config("druid.input_key_back", "key_back"))
+M.ACTION_ENTER = hash(sys.get_config("druid.input_key_enter", "key_enter"))
+M.ACTION_MULTITOUCH = hash(sys.get_config("druid.input_multitouch", "multitouch"))
+M.ACTION_BACKSPACE = hash(sys.get_config("druid.input_key_backspace", "key_backspace"))
 
-M.ACTION_TOUCH = hash("touch")
-M.ACTION_SCROLL_UP = hash("scroll_up")
-M.ACTION_MULTITOUCH = hash("multitouch")
-M.ACTION_SCROLL_DOWN = hash("scroll_down")
+M.ACTION_SCROLL_UP = hash(sys.get_config("druid.input_scroll_up", "scroll_up"))
+M.ACTION_SCROLL_DOWN = hash(sys.get_config("druid.input_scroll_down", "scroll_down"))
 
 
 M.RELEASED = "released"